DBGenie_schemaviz/sql-er-diagram-generator/requirements.txt
sqlparse
graphviz>=0.20.0
pydot
Flask==3.0.2
mysql-connector-python>=8.0.0